requirements.txt
# Core Robot Framework testing framework
robotframework==7.3.2

# Browser automation library for Robot Framework (uses Playwright)
robotframework-browser==19.6.2

# Parallel execution for Robot Framework
robotframework-pabot==4.3.2

# Environment variable loading from .env files
python-dotenv==1.0.1

# Google authentication and OAuth libraries for API testing
google-auth==2.38.0
google-auth-oauthlib==1.2.1

# HTML/XML parsing for web scraping and content validation
beautifulsoup4==4.13.3

# HTTP requests library for Robot Framework API testing
robotframework-requests==0.9.7

# =============================================================================
# Code Quality and Linting Tools
# =============================================================================
# Robot Framework linter and formatter
robotframework-robocop>=6.0.0

# Python linter
ruff>=0.1.0
